requirements.txt
signalwire-sdk>=3.0.1
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.32.0
gunicorn>=21.0.0
uvicorn>=0.34.2
